Update cache-settings for Reports page (#477)
The Reports page is not updating properly when you do stuff such as
removing a policy, or clicking back and forth between the vertical
navigation elements. Instead, we can use cache-and-network which returns
the cache elements first - and makes a network call in the backend,
which is only used to populate the data if it's different to what was in
the cache.