RHICOMPL-611 - Refreshing on details view should not redirect (#576)
* RHICOMPL-611 - Refreshing on details view should not redirect

Before this change, the app was listening to the APP_NAVIGATION event in
redux for all pages. This meant that if you refresh on any details page,
the APP_NAVIGATION event will retrigger and try to redirect to the base
URL for that details page (e.g: /scappolicies for /scappolicies/:policy_id)

Instead, we should only listen to the events we care about, which are
the ones for moving to the vertical navigation items. All other links
through the app, shouldn't have to go through this.

* Add routerParams to main pages

* Update snaps

* Update testing to exclude routers