RHICOMPL-556 - Run onWizardFinish only when button is clicked (#597)
Currently onWizardFinish runs automatically when the systems have been
assigned to the policy. This function is used mainly to reload the table
of SCAP policies.

Instead, we should only run it when the button is clicked - this allows
us to NOT automatically reload the page when the wizard finishes and we
have just created our **first** policy or a Policy through the Reports
page empty state.