[RHICOMPL-1036] Policy aware SystemStore reducer
The system reducer should pick profiles that are part of the policy.

This covers case when a policy report is shown (with internal profile
loaded) having a system scanned with a different SSG.  In that case the
host would have different profileId, however it is still part of the
policy.