fix(ReportDetails): RHICOMPL-1138 Show hosts with policy results
This changes the systems table to accept a defaultFilter prop which will
be added to user filters during the API request. This is to allow
specifying a new filter for the ReportDetails page to only show hosts
that have test results for a given policy (a set of profiles associated
to the policy). The new filter takes policy ID or any profile ID
associated to the policy.

Signed-off-by: Andrew Kofink <akofink@redhat.com>