fix(InventoryTable): RHICOMPL-1442 redux states
This change ensures the redux state stays set as intended. This includes
the rendered columns and rows as well as whether the table is loaded or
not.

It is very important to realize that the redux actions in our
SystemStore have (basically) no guarantee of any execution order. One
caveat is the *_PENDING and *_FULFILLED actions, which always happen in
that order.

This adds a GET_SYSTEMS_PENDING action to allow compliance to reset
columns and rows in addition to system information when our API
request begins. This is necessary to ensure old state is not persisted
into a new table, such as when viewing policy systems and then
immediately editing policy systems without refreshing.

Signed-off-by: Andrew Kofink <akofink@redhat.com>