fix(EditPolicy): RHICOMPL-1742 save button disabled on empty systems
The updatedPolicy has been set more than once in EditPolicyForm whithin
three different useEffect callbacks.  Because the effects did not have it
as a dependency (for triggering reasons) they do not get the updated
content from the previous useEffect.  Since these useEffects can fire
all at the same time, one would not get the update from the other.  This
fixes this by using the previous state to merge with the new one.

There are alternatives for fixing this, like splitting the
updatedPolicy to three separate state variables hold by the EditPolicy.