fix(TabbedRules): RHICOMPL-1782 canonical profiles with different OS
Fixes cases when the tabbed rules would get two or more canonical
profiles, where each would get different OS minor version set.  This can
be observed for RHEL 6 systems.

For this it had to change the defaultProfileId to defaultTab that must
include a profile id and an OS minor version.