fix(SystemsTable): Enable selection and refactor useSystemBulkSelect (#2201)
* fix(SystemsTable): Enable selection and refactor useSystemBulkSelect

Fixes https://issues.redhat.com/browse/RHINENG-13404.

This enables selection in the Systems table when the REST API feature flag is set on. This also revisits useSystemBulkSelect and makes it cache the items fetched in the table.

* fix(useSystemBulkSelect): Reset items on filters change

This adds a reset callback to useLoadedItems hook. The function is called by useSystemBulkSelect to clear the cache if active filters have changed for the correct "select all" behaviour.

* test: Add two new test for unionBy