fix(RHINENG-14334): System detail shows empty GridItem
There is the case when you have your system assigned to multiple policies but it's not reported against each policy - we will still try to show a card (you can see an GridItem - we show there an empty component).
This is happening because we have multiple reports returned but some of them have 0 test results ( as system is never reported)

To repro:
- Go to Compliance -> Systems
- Find a system with multiple policies but at least one policy with 0   test results (data.meta.total === 0)

Before, the GridItem would render, but with no card.

Now, the GridItem has been moved into the child component and will only display the GridItem if the meta.total > 0.